Put route to apis for consise app.ts
